# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Node API",
    "description": "Template API project for Node.js",
    "version": "1.0.0"
  },
  "basePath": "/api/node",
  "produces": [
    "application/json"
  ],
  "paths": {
    "/_checkAPIkey": {
      "get": {
        "operationId": "checkAPIkey",
        "summary": "GET a test endpoint that is protected by api key to validate that consumer has a proper api key",
        "description": "",
        "parameters": [],
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "An empty response"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": [
              "read"
            ]
          }
        ]
      }
    },
    "/v1/courseRoundStaitics/{roundEndDate}": {
      "post": {
        "operationId": "requestRoundStatisticsByLadokId",
        "summary": "Saves an object in the database",
        "description": "Saves changes to an object in the database",
        "parameters": [
          {
            "name": "roundEndDate",
            "in": "path",
            "description": "End date of the roun(s) to get the correct result, format yyyy-mm-dd",
            "required": true,
            "type": "string"
          },
          {
            "name": "Ladok roundId list",
            "in": "body",
            "description": "The data to save",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Data"
            }
          }
        ],
        "tags": [
          "v1"
        ],
        "responses": {
          "200": {
            "description": "A successful save"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [{
          "api_key": [
            "write"
          ]
        }]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header",
      "scopes": {
        "read": "Read access to data",
        "write": "Write access to data"
      }
    }
  },
  "definitions": {
    "Data": {
      "properties": {
        "roundLadokIdList": {
          "type": "array",
          "items": "string",
          "description": "Description of the data"
        }
      }
    },
    "Error": {
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    },
    "UserList": {
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    }
  }
}